requirements.txt
#
sqlparse==0.2.4
requests==2.21.0
igramscraper==0.3.5

Django>=2.2, <2.3
djangorestframework==3.11.1
djangorestframework-simplejwt>=4.4, <4.5
django-filter

python-decouple
celery>=4.4, <4.5
django-celery-beat
drf-yasg==1.17.1
coverage==5.1
mock==4.0.2
sentry-sdk==0.14.2
pycrypto==2.6.1
django-picklefield==3.0.1
PyJWT
psycopg2-binary
python-memcached
Pillow
zeep
django-tinymce
